requirements.txt
asgiref==3.2.3
Django==3.0.2
django-subdomains==2.1.0
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.0
